Changing Voters, utility, and ignorance values
In the design, it is clear for any real-world election the
number of Voters is going to have to be at least the number of
Candidates. While it is conceivable an election for public
office could have fewer Voters than Candidates, I am not aware
of any at this point in time. If there is a need to change this
in the future, I can still do so.

The change in the utility and ignorance values is because they
represent real-valued random variables whose distributions are
not known. The only question I still have is what to use as the
variance for distribution: the current traditional value of 1,
Gauss's value of 1/2, or Stigler's value of 1/(2π)? At this
point, I am inclined to use 1 because I see no reason at this
moment to depart from the traditional value. If someone presents
a case showing why 1 is inadequate, I can change later.