initializing Candidate utilities
For now, I am using a randomly selected value for both perceived
and actual utilities. This crude approach claims a justification
of "since the factors which influence Voter ignorance and how
much each factor influences that ignorance are not automatically
know, using the same pseudorandom number generator make sense;
what is good enough for determining the actually utility is good
enough for determining the perceived utility". Methods of
calculating perceived utility based on the actual utility can be
added later.