recording all Candidate and total societal utility value in multiple ways
Notes:

I want to adjust the code so not only the societally optimum
Candidate is recorded for each election but all Candidates and
Their respective societal utility values. I want to record both
the Candidates' societal utilities when utilities are weighted
by number of votes a Voter is permitted and the Candidates'
societal utilities when those utilities are weighted equally.
Such a record will allow future analysis for cases such as, "If
Legislators are given a number of votes equal to the number of
Constituents and the utility values are the average utility
values of said Constituents, which Candidate has the societally
optimum value to All Constituents?" and "If Shareholders are
given a number of votes in proportion to Their shares, which
Candidate has the societally optimum value to All Shareholders?"
The former case necessitates determining the Candidate by weight
with votes weighted during the simulations and the latter
necessitates determining the Candidate by equal weight with
votes weighted during the simulations.
======================================================================
Model:

I will collect Each Candidates' societal utility values, both
weighted and unweighted, into two respective collections and
record copies of those collections in the respective `Election`
object.

I can later sort these collections in order to determine the
societally optimum Candidate as measured by the collections'
respective weighting methods.
======================================================================
Analysis:

I will need to add two collections to the `Election` type, one
representing the weighted results and the other representing the
unweighted results.
======================================================================